server_code/requirements.txt
openai==1.3.5
tavily-python==0.2.8